# swagger.json
{
	"basePath": "/api/v1",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"StatusType": {
			"enum": [
				"Valide",
				"Invalide",
				"Undefined"
			],
			"type": "string"
		},
		"BusinessStatusModel": {
			"properties": {
				"name": {
					"type": "string"
				},
				"rbqNum": {
					"type": "string"
				},
				"status": {
					"$ref": "#/definitions/StatusType"
				},
				"isValid": {
					"type": "boolean"
				},
				"neqId": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"name",
				"rbqNum",
				"status",
				"isValid",
				"neqId"
			],
			"type": "object",
			"additionalProperties": true
		},
		"BusinessInfoModel": {
			"properties": {
				"name": {
					"type": "string"
				},
				"otherName": {
					"type": "string"
				},
				"rbqNum": {
					"type": "string"
				},
				"status": {
					"$ref": "#/definitions/StatusType"
				},
				"deliveranceDate": {
					"type": "string",
					"format": "date-time"
				},
				"paymentDate": {
					"type": "string",
					"format": "date-time"
				},
				"neqId": {
					"type": "number",
					"format": "double"
				},
				"address": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"telephone": {
					"type": "string"
				}
			},
			"required": [
				"name",
				"otherName",
				"rbqNum",
				"status",
				"deliveranceDate",
				"paymentDate",
				"neqId",
				"address",
				"email",
				"telephone"
			],
			"type": "object",
			"additionalProperties": true
		}
	},
	"info": {
		"title": "mender-validation-gateway",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/license/rbq/status/{rbqNum}": {
			"get": {
				"operationId": "GetStatusByRbq",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/BusinessStatusModel"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "rbqNum",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/license/rbq/{rbqNum}": {
			"get": {
				"operationId": "GetInfoByRbq",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/BusinessInfoModel"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "rbqNum",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/license/neq/status/{neqId}": {
			"get": {
				"operationId": "GetStatusByNeq",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/BusinessStatusModel"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "neqId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/license/neq/{neqId}": {
			"get": {
				"operationId": "GetInfoByNeq",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/BusinessInfoModel"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "neqId",
						"required": true,
						"type": "string"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {},
	"host": "localhost:3000"
}